news_pulse_app/requirements.txt
streamlit
requests
python-dotenv
passlib[bcrypt]
nltk
pandas

# Added for Gemini support
google-genai
